Added interaction-tracking/subscriptions bundle and split tests
