+++ packages/interaction-tracking/package.json
@@ -7,6 +7,7 @@
     "LICENSE",
     "README.md",
     "index.js",
+    "subscriptions.js",
     "cjs/",
     "umd/"
   ],
